MDEV-37681 SIGSEGV on TRUNCATE GTT after failed RENAME under LOCK TABLE
WHILE RENAME was covered in the reference counting feature,
ALTER TABLE REMANE was missed, as it's turned to be a different code
path.

Exit earlier in mysql_alter_table, if global temporary table is opened
in this connection, or in the other connection. Add timeout argument to
simple_rename_or_index_change, which can be 0 to conform Oracle behavior

Parent issue:
MDEV-35915 Implement Global temporary tables