MDEV-37850 Wrong error on DROP TABLE GTT after DROP DATABASE
Turns out that temporary tables are not dropped as part of DROP
DATABASE.

This patch implements dropping child global temporary tables for the
currently dropping database.

That is, if the child table was opened in this connection, it will be
dropped together with the parent table.

If another connection attempts to drop the database, it will wait while
the table is in use.

Parent issue:
MDEV-35915 Implement Global temporary tables